Change Updater | 0.1.4

+++ babysploit/updater.py
@@ -1,27 +1,16 @@
 def checkupdate():
-    import requests
+    import requests, os, subprocess
+    from tqdm import tqdm, trange
     with open("babysploit/version", "r") as fwv:
         data = fwv.read().replace(" ", "")
     cv = requests.get("https://raw.githubusercontent.com/M4cs/BabySploit/master/babysploit/version").text.replace(" ", "")
     if data == cv:
-        state = "uptodate"
-    elif data != cv:
-        state = "outofdate"
-    return state, data, cv
-
-def update():
-    import os, subprocess
-    ans = os.path.exists(".git")
-    if ans == True:
-        print("Starting Update...")
         pass
-    elif ans == False:
-        print("[!] Error Updating. You Did Not Clone The Repository. [!]")
-        exit()
-    print("Fetching New Version...")
-    subprocess.check_output("git fetch --all", shell=True)
-    print("Updating...")
-    subprocess.check_output("git reset --hard origin/master", shell=True)
-    os.system("pip3 install -r requirements.txt")
-    print("Exiting Updater...")
-    exit()
+    elif data != cv:
+        print("[!] Update Found | Version: %s [!]" % cv)
+        ask = str(input("[y\\n] ").lower())
+        if ask == "y":
+            for i in tqdm(range(int(100)), desc="Updating"):
+                subprocess.check_output("git fetch --all", shell=True)
+                subprocess.check_output("git reset --hard origin/master", shell=True)
+                subprocess.check_output("pip3 install -r requirements.txt", shell=True)
